moving configuration to a folder
